requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
snowflake-sqlalchemy==1.5.1
snowflake-connector-python
werkzeug==3.0.1
gunicorn==21.2.0
